+++ src/main/java/com/dio/santander/bankline/api/service/CorrentistaService.java
@@ -7,6 +7,7 @@ import com.dio.santander.bankline.api.repository.CorrentistaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Service
@@ -18,7 +19,7 @@ public class CorrentistaService {
     correntista.setCpf(novoCorrentista.getCpf());
     correntista.setNome(novoCorrentista.getNome());
     Conta conta = new Conta();
-    conta.setSaldo(0.0);
+    conta.setSaldo(BigDecimal.valueOf(0));
     conta.setNumero(new Date().getTime());
     correntista.setConta(conta);
     repository.save(correntista);
